Move imgaes from img folder to assets and delete it + small changes to redux setup
