+++ client/package.json
@@ -20,6 +20,7 @@
     "react-test-renderer": "^16.6.3",
     "react-validation": "^3.0.7",
     "redux": "^4.0.1",
+    "redux-devtools-extension": "^2.13.8",
     "redux-form": "^8.2.0",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
