+++ client/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 
 import App from "./Components/App";
-import store from "./redux/store";
+import store from "./store";
 
 import "./index.css";
 import "./normalize.css";
