--- client/src/redux/actions/types.js
@@ -1,1 +1,0 @@
-.gitkeep
